expections

+++ backend/src/main/java/com/autoflex/projedata/service/RawMaterialService.java
@@ -5,16 +5,21 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 
 import com.autoflex.projedata.entity.RawMaterial;
-import com.autoflex.projedata.exception.DuplicateProductCodeException;
+import com.autoflex.projedata.exception.DuplicateRawMaterialCodeException;
+import com.autoflex.projedata.exception.LinkedMaterialRawProductException;
+import com.autoflex.projedata.repository.ProductRawMaterialRepository;
 import com.autoflex.projedata.repository.RawMaterialRepository;
 
 @Service
 public class RawMaterialService {
 
     private final RawMaterialRepository repository;
+    private final ProductRawMaterialRepository productRawMaterialRepository;
 
-    public RawMaterialService(RawMaterialRepository repository) {
+    public RawMaterialService(RawMaterialRepository repository,
+            ProductRawMaterialRepository productRawMaterialRepository) {
         this.repository = repository;
+        this.productRawMaterialRepository = productRawMaterialRepository;
     }
 
     public List<RawMaterial> findAll() {
@@ -28,7 +33,7 @@ public class RawMaterialService {
 
     public RawMaterial save(RawMaterial rawMaterial) {
         if (repository.existsByCode(rawMaterial.getCode())) {
-            throw new DuplicateProductCodeException("Raw material code already exists");
+            throw new DuplicateRawMaterialCodeException("Raw material code already exists");
         }
         return repository.save(rawMaterial);
     }
@@ -37,7 +42,7 @@ public class RawMaterialService {
         RawMaterial rm = findById(id);
 
         if (!rm.getCode().equals(updated.getCode()) && repository.existsByCode(updated.getCode())) {
-            throw new DuplicateProductCodeException("Raw material code already exists");
+            throw new DuplicateRawMaterialCodeException("Raw material code already exists");
         }
         rm.setCode(updated.getCode());
         rm.setName(updated.getName());
@@ -46,6 +51,9 @@ public class RawMaterialService {
     }
 
     public void delete(Long id) {
+        if (productRawMaterialRepository.existsByIdRawMaterialId(id)) {
+            throw new LinkedMaterialRawProductException("Cannot delete raw material linked to products");
+        }
         repository.deleteById(id);
     }
 }
